Add function to check drone connection with navdata
Previously we had planned to use the ping and pingAll methods to check if  drones were connected but using navdata removes the need to use an extra library and decreases the network strain that comes with failed pings.

This commit removes those functions and updates the status object and function with data used for testing.

It also changes the allConnected function to a get method, because it's easier to type that way.