Reconnection for lost connection, not power loss
Upon timeout the navdata event listeners on drones were being removed, despite
the fact that the event listeners contained the re-connection functionality.
The removal of the event listeners has now been removed. This allowed drones
to which a connection has been lost (for instance the computer running the
server disconnects from the network) to reconnect upon re-establishment of the
connection, but for reasons which are currently unknown, this does not allow
reconnection of drones that have restarted.

Additionally, drones.add was not running through for drones that had already
been stored, meaning that, for one, the lights would not flash again. It was
changed to run through (with the already existant drone), even if the drone
with that ID did already exist.

It is worth noting that throughout this commit message, reconnection refers to
the drone beginning sending navdata again